Refactor payment관련 파일 위치 이동
